refactor: 주문 생성 리팩토링
